Add PostgreSQL, SEQ, and monitoring services
This update introduces new services and configurations for
PostgreSQL, SEQ, Prometheus, and Grafana in the `Program.cs` file.
A password parameter for SQL Server has been added, and the handling
of volumes and bind mounts for persistent data storage has been
adjusted. The application is now set to wait for these new services
to be ready before proceeding, ensuring a smoother startup process.

Co-authored-by: Abel Briones Ramírez <abel.briones@exxerpro.com>